Update deps and create rust workspace for backend
The rust code now lives in the backend/ sub folder. The root project is
a cargo workspace now.

Updated deps, notably ft-sdk and rust-toolchain to latest nightly.
Some TODOs have been added to make the code compile for now. Running
this code might not produce the same result as it used (email sending
logic is not tested and might require some changes in user's fastn package)